Opsætning af Kalibrering
udvikling af udregning og tilgang til data

+++ DTO_blodtryksmåler/DTO_kalibrer.cs
@@ -9,10 +9,10 @@ namespace DTO_blodtryksmåler
     [Serializable]
     public class DTO_kalibrer
     {
-        public int Måling1 { get; set; }
-        public  int Måling2 { get; set; }
-        public  int Read1 { get; set; }
-        public  int Read2 { get; set; }
+        public double   Måling1 { get; set; }
+        public  double Måling2 { get; set; }
+        public  double Read1 { get; set; }
+        public  double Read2 { get; set; }
         public double Factor { get; set; }
     }
 }
